+++ autots/tools/fast_kalman.py
@@ -99,537 +99,8 @@ form 1 to 3 as long as they are reasonable from the point of view of matrix
 multiplication and numpy broadcasting rules. Matrix operations are applied on
 the *last* two axes of the arrays.
 """
-import random
 import numpy as np
 from functools import wraps
-
-
-def random_state_space_original():
-    """Return randomly generated statespace models."""
-    n_dims = random.choices([1, 2, 3, 4, 8], [0.1, 0.2, 0.3, 0.4, 0.3])[0]
-    if n_dims == 1:
-        st = np.array([[1]])
-        obsmod = np.random.randint(1, 3, (1, n_dims))
-        procnois = np.diag(np.random.exponential(0.01, size=(n_dims)).round(3))
-    else:
-        st = np.random.choice([0, 1, -1], p=[0.75, 0.2, 0.05], size=(n_dims, n_dims))
-        st[0, 0] = 1
-        st[0, -1] = 0
-        if n_dims == 2:
-            obsmod = np.array([[1, 0]])
-        else:
-            obsmod = np.array([[1, 1] + [0] * (n_dims - 2)])
-        procnois = (
-            np.diag([0.2 / random.choice([1, 5, 10]), 0.001] + [0] * (n_dims - 2)) ** 2
-        ).round(3)
-    obsnois = random.choices(
-        [1.0, 10.0, 2.0, 0.5, 0.2, 0.05, 0.001],
-        [0.8, 0.05, 0.05, 0.05, 0.05, 0.05, 0.05],
-    )[0]
-    while (
-        st.size > 1
-        and np.all(st == 1)
-        # or (st.size > 3 and np.isin(st.sum(axis=1), [0]).any())
-    ):
-        st, procnois, obsmod, obsnois = random_state_space()
-    return st, procnois, obsmod, obsnois
-
-
-def ensure_stability(st):
-    eigenvalues, eigenvectors = np.linalg.eig(st)
-    # Scale eigenvalues to ensure their absolute values are less than 1
-    stable_eigenvalues = eigenvalues / (np.abs(eigenvalues) + 1e-5)
-    st_stable = eigenvectors @ np.diag(stable_eigenvalues) @ np.linalg.inv(eigenvectors)
-    return st_stable.real
-
-
-def random_matrix(rows, cols, density=0.2):
-    matrix = np.random.randn(rows, cols)
-    sparsity_mask = np.random.rand(rows, cols) < density
-    return np.where(sparsity_mask, matrix, 0)
-
-
-def random_state_space(tries=15):
-    for _ in range(tries):
-        try:
-            n_dims = random.choices([1, 2, 3, 4, 8], [0.1, 0.2, 0.3, 0.4, 0.3])[0]
-            st = random_matrix(n_dims, n_dims, density=0.5)
-            st = ensure_stability(st)
-            obsmod = random_matrix(
-                1, n_dims, density=1.0
-            )  # Full observation for simplicity
-            procnois = np.diag(np.random.exponential(0.01, size=n_dims)).round(3)
-            obsnois = np.random.exponential(1.0)
-
-            if np.all(np.abs(np.linalg.eigvals(st)) < 1):  # Check stability
-                return st, procnois, obsmod, obsnois
-        except Exception:
-            pass
-    # fallback
-    return random_state_space_original()
-
-
-def holt_winters_damped_matrices(M, alpha, beta, gamma, phi=1.0):
-    """Not sure if this is correct. It's close, at least."""
-    # State Transition Matrix F
-    # Level & Trend Equations
-
-    F_lt = np.array(
-        [  # not sure about having the alpha and beta here
-            [1 + (alpha * (1 - phi)), phi],
-            [beta * (1 - phi), phi],
-        ]
-    )
-    # Seasonal Equation
-    F_s = np.eye(M, M, -1)  # This creates an identity matrix and shifts it down
-    first_row = np.zeros((1, M))
-    first_row[0, -1] = 1
-    F_s = np.vstack([first_row, F_s[:-1]])
-    F_top = np.hstack([F_lt, np.zeros((2, M))])
-    F_bottom = np.hstack([np.zeros((M, 2)), F_s])
-    F = np.vstack([F_top, F_bottom])
-    """
-    F = np.zeros((M+2, M+2))
-    F[0, 0] = 1
-    F[0, 1] = phi
-    F[1, 0] = 1
-    F[1, 1] = phi
-    for i in range(2, M+2):
-        F[i, i] = 1
-    """
-
-    # Process Noise Covariance Q
-    Q = np.zeros((M + 2, M + 2))
-    # Assuming the same variance for all states for simplicity.
-    # Modify these values based on specific requirements.
-    Q[0, 0] = alpha
-    Q[1, 1] = beta
-    for i in range(2, M + 2):
-        Q[i, i] = gamma
-
-    # Observation Matrix H
-    H = np.zeros((1, M + 2))
-    H[0, 0] = 1
-    H[0, 1] = 1
-    H[0, -1] = 1
-
-    # Observation Noise Covariance R
-    R = np.array([[1]])
-
-    return F, Q, H, R
-
-
-def new_kalman_params(method=None, allow_auto=True):
-    if method in ['fast']:
-        em_iter = random.choices([None, 10], [0.8, 0.2])[0]
-    elif method == "superfast":
-        em_iter = None
-    elif method == "deep":
-        em_iter = random.choices([None, 10, 20, 50, 100], [0.3, 0.6, 0.1, 0.1, 0.1])[0]
-    else:
-        em_iter = random.choices([None, 10, 30], [0.3, 0.7, 0.1])[0]
-
-    K = random.choices([2, 3, 6, 8], [0.1, 0.6, 0.1, 0.1])[0]
-    sigma_level2 = 1e-2  # Placeholder
-    sigma_slope2 = 1e-3  # Placeholder
-    sigma_weekly2 = 1e-2  # Placeholder
-    sigma_fourier2 = 1e-2  # Placeholder
-
-    Q = np.block(
-        [
-            [sigma_level2, 0, np.zeros((1, 6)), np.zeros((1, 2 * K))],
-            [0, sigma_slope2, np.zeros((1, 6)), np.zeros((1, 2 * K))],
-            [np.zeros((6, 2)), sigma_weekly2 * np.eye(6), np.zeros((6, 2 * K))],
-            [
-                np.zeros((2 * K, 2)),
-                np.zeros((2 * K, 6)),
-                sigma_fourier2 * np.eye(2 * K),
-            ],
-        ]
-    )
-
-    # Deterministic Trend
-    deterministic_trend = np.array([[1, 1], [0, 1]])
-
-    weekly = np.eye(6, k=1)
-    weekly[-1, 0] = -1  # Link the last state to the first to maintain cyclicality
-
-    harmonics = []
-    for k in range(1, K + 1):
-        angle = 2 * np.pi * k / 365.25
-        matrix = np.array(
-            [[np.cos(angle), -np.sin(angle)], [np.sin(angle), np.cos(angle)]]
-        )
-        harmonics.append(matrix)
-
-    # Stack Fourier matrices
-    fourier_yearly = np.block([harmonics[i] for i in range(K)])
-
-    # The top part
-    # F_top is of shape (2, 2+6+2K)
-    F_top = np.hstack(
-        [
-            deterministic_trend,  # shape: (2, 2)
-            np.zeros((2, 6)),  # shape: (2, 6)
-            np.zeros((2, 2 * K)),  # shape: (2, 2K)
-        ]
-    )
-
-    # The middle part (weekly seasonality)
-    # F_mid is of shape (6, 2+6+2K)
-    F_mid = np.hstack(
-        [
-            np.zeros((6, 2)),  # shape: (6, 2)
-            weekly,  # shape: (6, 6)
-            np.zeros((6, 2 * K)),  # shape: (6, 2K)
-        ]
-    )
-
-    # The bottom part (yearly Fourier seasonality)
-    # F_bot is of shape (2K, 2+6+2K)
-    F_bot = np.hstack(
-        [
-            np.zeros((2 * K, 2)),  # shape: (2K, 2)
-            np.zeros((2 * K, 6)),  # shape: (2K, 6)
-            fourier_yearly.T,  # shape: (2K, 2K)
-            np.zeros((2 * K, K * 2 - 2)),
-        ]
-    )
-    # Vertically stack
-    F = np.vstack([F_top, F_mid, F_bot])
-    # Generating H for Fourier terms
-    H_fourier = []
-    for k in range(1, K + 1):
-        H_fourier.extend(
-            [np.cos(2 * np.pi * k / 365.25), np.sin(2 * np.pi * k / 365.25)]
-        )
-    H = np.hstack(([1, 0], np.ones(6), H_fourier))
-
-    F_hw, Q_hw, H_hw, R_hw = holt_winters_damped_matrices(
-        M=7,
-        alpha=random.choice([0.9, 1.0, 0, 0.3]),
-        beta=random.choice([0.9, 1.0, 0, 0.3]),
-        gamma=random.choice([0.9, 1.0, 0, 0.3]),
-        phi=random.choice([0.9, 1.0, 0.995, 0.98]),
-    )
-
-    params = random.choices(
-        # the same model can sometimes be defined in various matrix forms
-        [
-            # floats are phi
-            'ets_aan',
-            {
-                'model_name': 'local linear stochastic seasonal dummy',
-                'state_transition': [
-                    [1, 0, 0, 0],
-                    [0, -1, -1, -1],
-                    [0, 1, 0, 0],
-                    [0, 0, 1, 0],
-                ],
-                'process_noise': [
-                    [1, 0, 0, 0],
-                    [0, 1, 0, 0],
-                    [0, 0, 0, 0],
-                    [0, 0, 0, 0],
-                ],
-                'observation_model': [[1, 1, 0, 0]],
-                'observation_noise': random.choices(
-                    [0.25, 0.04, 0.57, 1.0, 0.08, 0.02, 0.8, 'auto'],
-                    [0.2, 0.3, 0.3, 0.3, 0.1, 0.05, 0.05, 0.025],
-                )[
-                    0
-                ],  # 'auto' is slow
-            },
-            {
-                'model_name': 'local linear stochastic seasonal 7',
-                'state_transition': [
-                    [1, 1, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0],
-                    [0, 1, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0],
-                    [0.0, 0, -1.0, -1.0, -1.0, -1.0, -1.0, 0.0],
-                    [0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0],
-                    [0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0],
-                    [0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0],
-                    [0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0],
-                    [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0],
-                ],
-                'process_noise': [
-                    [1, 0, 0, 0, 0, 0, 0, 0],
-                    [0, 1, 0, 0, 0, 0, 0, 0],
-                    [0, 0, 1, 0, 0, 0, 0, 0],
-                    [0, 0, 0, 1, 0, 0, 0, 0],
-                    [0, 0, 0, 0, 1, 0, 0, 0],
-                    [0, 0, 0, 0, 0, 1, 0, 0],
-                    [0, 0, 0, 0, 0, 0, 0, 0],
-                    [0, 0, 0, 0, 0, 0, 0, 0],
-                ],
-                'observation_model': [[1, 0, 1, 0, 0, 0, 0, 0]],
-                'observation_noise': random.choice([0.25, 'auto']),
-            },
-            {
-                'model_name': 'MA',
-                'state_transition': [[1, 0], [1, 0]],
-                'process_noise': [[0.2, 0.0], [0.0, 0]],
-                'observation_model': [[1, 0.1]],
-                'observation_noise': 1.0,
-            },
-            {
-                'model_name': 'AR(2)',
-                'state_transition': [[1, 1], [0.1, 0]],
-                'process_noise': [[1, 0], [0, 0]],
-                'observation_model': [[1, 0]],
-                'observation_noise': 1.0,
-            },
-            {
-                'model_name': 'ucm_deterministic_trend',
-                'state_transition': [[1, 1], [0, 1]],
-                'process_noise': [[0.01, 0], [0, 0.01]],  # these would be tuned
-                'observation_model': [[1, 0]],
-                'observation_noise': 0.1,  # this would be tuned
-            },
-            {
-                'model_name': 'X1',
-                'state_transition': [[1, 1, 0], [0, 1, 0], [0, 0, 1]],
-                'process_noise': [
-                    [0.1, 0.0, 0.0],
-                    [0.0, 0.01, 0.0],
-                    [0.0, 0.0, 0.1],
-                ],
-                'observation_model': [[1, 1, 1]],
-                'observation_noise': 1.0,
-            },
-            # I believe this is a seasonal ETS model but I would like confirmation on that
-            {
-                'model_name': "local linear hidden state with seasonal 7",
-                'state_transition': [
-                    [1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0],
-                    [0.0, -1.0, -1.0, -1.0, -1.0, -1.0, -1.0, 0.0],
-                    [0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0],
-                    [0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0],
-                    [0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0],
-                    [0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0],
-                    [0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0],
-                    [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0],
-                ],
-                'process_noise': [
-                    [0.0016, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0],
-                    [0.0, 1e-06, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0],
-                    [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0],
-                    [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0],
-                    [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0],
-                    [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0],
-                    [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0],
-                    [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0],
-                ],
-                'observation_model': [[1, 1, 0, 0, 0, 0, 0, 0]],
-                'observation_noise': random.choice(
-                    [0.25, 0.5, 1.0, 0.04, 0.02, 'auto']
-                ),
-            },
-            {
-                'model_name': "factor",
-                'state_transition': [
-                    [1, 1, 0, 0, 0, 0],
-                    [0, 1, 0, 0, 0, 0],
-                    [0, 0, 1, 0, 0, 0],
-                    [0, 0, 0, 1, 1, 0],
-                    [0, 0, 0, 0, 1, 0],
-                    [0, 0, 0, 0, 0, 1],
-                ],
-                'process_noise': [
-                    [1, 0, 0, 0, 0, 0],
-                    [0, 1, 0, 0, 0, 0],
-                    [0, 0, 0, 0, 0, 0],
-                    [0, 0, 1, 0, 0, 0],
-                    [0, 0, 0, 1, 0, 0],
-                    [0, 0, 0, 0, 0, 0],
-                ],
-                'observation_model': [[1, 0, 0, 0, 0, 0]],
-                'observation_noise': 0.04,
-            },
-            {
-                'model_name': "ucm_deterministictrend_seasonal7",
-                'state_transition': [
-                    [1, 1, 0, 0, 0, 0, 0, 0],
-                    [0, 1, 0, 0, 0, 0, 0, 0],
-                    [0, 0, 1, 0, 0, 0, 0, 0],
-                    [0, 0, 0, 1, 0, 0, 0, 0],
-                    [0, 0, 0, 0, 1, 0, 0, 0],
-                    [0, 0, 0, 0, 0, 1, 0, 0],
-                    [0, 0, 0, 0, 0, 0, 1, 0],
-                    [0, 0, -1, -1, -1, -1, -1, -1],
-                ],
-                'process_noise': [
-                    [0.001, 0, 0, 0, 0, 0, 0, 0],
-                    [0, 0.001, 0, 0, 0, 0, 0, 0],
-                    [0, 0, 0.001, 0, 0, 0, 0, 0],
-                    [0, 0, 0, 0.001, 0, 0, 0, 0],
-                    [0, 0, 0, 0, 0.001, 0, 0, 0],
-                    [0, 0, 0, 0, 0, 0.001, 0, 0],
-                    [0, 0, 0, 0, 0, 0, 0.001, 0],
-                    [0, 0, 0, 0, 0, 0, 0, 0],
-                ],
-                'observation_model': [[1, 0, 1, 1, 1, 1, 1, 1]],
-                'observation_noise': random.choice([0.03, 'auto']),
-            },
-            "random",
-            364,
-            12,
-            {
-                'model_name': 'spline',
-                'state_transition': [[2, -1], [1, 0]],
-                'process_noise': [[1, 0], [0, 0]],
-                'observation_model': [[1, 0]],
-                'observation_noise': random.choice([0.1, 'auto']),
-            },
-            {
-                'model_name': 'locallinear_weekly_fourier',
-                'state_transition': F.tolist(),
-                'process_noise': Q.tolist(),
-                'observation_model': H[np.newaxis, ...].tolist(),
-                'observation_noise': random.choice([0.01, 0.1, 'auto']),
-            },
-            {
-                'model_name': 'holt_winters_damped',
-                'state_transition': F_hw.tolist(),
-                'process_noise': Q_hw.tolist(),
-                'observation_model': H_hw.tolist(),
-                'observation_noise': random.choice([0.01, 0.1, 1, 'auto']),
-            },
-            "dynamic_linear",
-            "random_original",
-        ],
-        [
-            0.1,
-            0.1,
-            0.1,
-            0.1,
-            0.1,
-            0.1,
-            0.1,
-            0.1,
-            0.1,
-            0.2,
-            0.1,
-            0.1,
-            0.1,
-            0.1,
-            0.1,
-            0.1,
-            0.1,
-            0.1,
-        ],
-    )[0]
-    if params in [364] and method not in ['deep']:
-        params = 7
-    if params == "random":
-        st, procnois, obsmod, obsnois = random_state_space()
-        params = {
-            'model_name': 'randomly generated',
-            'state_transition': st.tolist(),
-            'process_noise': procnois.tolist(),
-            'observation_model': obsmod.tolist(),
-            'observation_noise': obsnois,
-        }
-    elif params == "random_original":
-        st, procnois, obsmod, obsnois = random_state_space_original()
-        params = {
-            'model_name': 'randomly generated_original',
-            'state_transition': st.tolist(),
-            'process_noise': procnois.tolist(),
-            'observation_model': obsmod.tolist(),
-            'observation_noise': obsnois,
-        }
-    elif params == "dynamic_linear":
-        choices = [
-            0.0,
-            0.1,
-            0.2,
-            0.3,
-            0.4,
-            0.5,
-            0.6,
-            0.7,
-            0.8,
-            0.9,
-            1.0,
-            1.1,
-            1.2,
-            1.3,
-            1.4,
-            1.5,
-        ]
-        params = {
-            'model_name': 'dynamic linear',
-            'state_transition': [
-                [1, 1, 0, 0],
-                [0, 1, 0, 0],
-                [0, 0, random.choice(choices), 1],
-                [0, 0, random.choice(choices), 0],
-            ],
-            'process_noise': [
-                [random.choice(choices), 0, 0, 0],
-                [0, random.choice(choices), 0, 0],
-                [0, 0, random.choice(choices), 0],
-                [0, 0, 0, 0],
-            ],
-            'observation_model': [[1, 0, 1, 0]],
-            'observation_noise': 0.25,
-            'em_iter': 10,
-        }
-    elif params == "ets_aan":
-        choices = [
-            0.0,
-            0.01,
-            0.1,
-            0.2,
-            0.3,
-            0.4,
-            0.5,
-            0.6,
-            0.7,
-            0.8,
-            0.9,
-            1.0,
-            1.1,
-            1.2,
-            1.3,
-            1.4,
-            1.5,
-        ]
-        params = {
-            'model_name': 'local_linear_trend_ets_aan',
-            'state_transition': [[1, 1], [0, 1]],
-            'process_noise': [
-                [random.choice(choices), 0.0],
-                [0.0, random.choice(choices)],
-            ],
-            'observation_model': [[1, 0]],
-            'observation_noise': random.choice([0.25, 0.5, 1.0, 0.05]),
-        }
-    elif isinstance(params, int):
-        state_transition = np.zeros((params + 1, params + 1))
-        state_transition[0, 0] = 1
-        state_transition[1, 1:-1] = [-1.0] * (params - 1)
-        state_transition[2:, 1:-1] = np.eye(params - 1)
-        observation_model = [[1, 1] + [0] * (params - 1)]
-        level_noise = 0.2 / random.choice([0.2, 0.5, 1, 5, 10, 200])
-        season_noise = random.choice([1e-4, 1e-3, 1e-2, 1e-1])
-        process_noise_cov = (
-            np.diag([level_noise, season_noise] + [0] * (params - 1)) ** 2
-        )
-        params = {
-            'model_name': f'local linear hidden state with seasonal {params}',
-            'state_transition': state_transition.tolist(),
-            'process_noise': process_noise_cov.tolist(),
-            'observation_model': observation_model,
-            'observation_noise': 0.04,
-        }
-    params['em_iter'] = em_iter
-    if not allow_auto:
-        if params['observation_noise'] == 'auto':
-            params['observation_noise'] = 0.1
-    return params
-
 
 class Gaussian:
     def __init__(self, mean, cov):
